backend/requirements.txt
# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic[email]==2.5.2

# Database
sqlalchemy==2.0.23

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Selenium - Web Test
selenium==4.15.2
webdriver-manager==4.0.1

# Appium - Mobile Test
Appium-Python-Client==3.1.0

# HTTP
requests==2.31.0

# Excel
openpyxl==3.1.2
pandas==2.1.3
